+++ test/identities.js
@@ -1,11 +1,11 @@
 var chai = require('chai');
 chai.Should();
-chai.use(require('../index.js').chaiModule);
+chai.use(require('../dist/').chaiModule);
 var _ = require('lodash')
 
 describe('Identies in the full tree', function() {
   it("Different identities are valid", function() {
-    require('./data/identities.json').should.be.validSignalK;
+    require('./data/identities.json').should.be.validSignalKIgnoringSelf;
   });
 });