+++ test/racing.js
@@ -1,9 +1,9 @@
 var chai = require('chai');
 chai.Should();
-chai.use(require('../index.js').chaiModule);
+chai.use(require('../dist/').chaiModule);
 
 describe('Racing items ok in navigation', function() {
   it("Sample racing data is valid", function() {
-    require('./data/racing.json').should.be.validSignalK;
+    require('./data/racing.json').should.be.validSignalKIgnoringSelf;
   });
 });