+++ test/sails.js
@@ -1,6 +1,6 @@
 var chai = require('chai');
 chai.Should();
-chai.use(require('../index.js').chaiModule);
+chai.use(require('../dist/').chaiModule);
 
 describe('Sails', function() {
   it("sample tree is valid", function() {