+++ test/sources.js
@@ -1,11 +1,11 @@
 const chai = require('chai');
 const should = chai.should()
-chai.use(require('../index.js').chaiModule);
+chai.use(require('../dist/').chaiModule);
 const FullSignalK = require('../src/fullsignalk')
 
 describe('Sources in the full tree', function() {
   it("Sample full tree is valid", function() {
-    require('./data/sources.json').should.be.validSignalK;
+    require('./data/sources.json').should.be.validSignalKIgnoringSelf;
   });
 });
 
@@ -144,7 +144,7 @@ describe('Bad sources in delta', function() {
 
 describe('Multiple sources for the same path:', function() {
   it("value + values are valid", function() {
-    require('./data/multiple-values.json').should.be.validSignalK
+    require('./data/multiple-values.json').should.be.validSignalKIgnoringSelf
   });
 });