+++ test/trip.js
@@ -1,21 +1,21 @@
 var chai = require('chai');
 chai.Should();
-chai.use(require('../index.js').chaiModule);
+chai.use(require('../dist/').chaiModule);
 
 describe('Trip data is valid', function() {
   it("full tree is valid", function() {
-    require('./data/trip.json').should.be.validSignalK;
+    require('./data/trip.json').should.be.validSignalKIgnoringSelf;
   });
 });
 
 describe('Trip data is invalid - string value', function() {
   it("full tree is valid", function() {
-    require('./data/trip-bad1.json').should.not.be.validSignalK;
+    require('./data/trip-bad1.json').should.not.be.validSignalKIgnoringSelf;
   });
 });
 
 describe('Trip data is invalid-bad key', function() {
   it("full tree is valid", function() {
-    require('./data/trip-bad2.json').should.not.be.validSignalK;
+    require('./data/trip-bad2.json').should.not.be.validSignalKIgnoringSelf;
   });
 });