+++ test/vessel-time.js
@@ -1,9 +1,9 @@
 var chai = require('chai');
 chai.Should();
-chai.use(require('../index.js').chaiModule);
+chai.use(require('../dist/').chaiModule);
 
 describe('Time in the full tree', function() {
   it("Sample full tree is valid", function() {
-    require('../samples/vessel-time.json').should.be.validSignalK;
+    require('../samples/vessel-time.json').should.be.validSignalKIgnoringSelf;
   });
 });